[ADDED] pack icon
Former-commit-id: c4fc7aaef1306a3465ce09aa91b06652867520c5

+++ simp/include/simp/NodeShape.h
@@ -19,7 +19,7 @@ public:
 
 	uint32_t color;
 
-	uint32_t vertices_n;
+	uint16_t vertices_n;
 	uint16_t vertices[1];
 
 public:
